requirements.txt
tensorflow==2.1.0
pillow==8.3.2
numpy==1.19.3
streamlit==1.12.0
streamlit-authenticator==0.2.1
deta==1.1.0
python-dotenv==0.21.0
pandas==1.3.5
pip==23.0
